Cast segmentation data to uint8 before processing
Modified the segmentation data casting to uint8 for compatibility with downstream processes in the preview_jpg function. This ensures that the data is in the correct format for subsequent transformations and reduces potential type-related errors.

+++ totalspineseg/utils/preview_jpg.py
@@ -242,7 +242,7 @@ def _preview_jpg(
             seg.affine
         except:
             seg = freesurfer.load(seg_path)
-            seg = tio.LabelMap(tensor=np.asanyarray(seg.dataobj)[None, ...], affine=seg.affine)
+            seg = tio.LabelMap(tensor=np.asanyarray(seg.dataobj).astype(np.uint8)[None, ...], affine=seg.affine)
         seg = tio.ToCanonical()(seg)
         seg = tio.Resample(image)(seg)
